Remove "Default License" concept
This concept was created to ensure that the specific Underlying License
being exercised was unambiguous and at the same time save licensees
the trouble of specifying an Underlying License in most cases.

However, this ambiguity is mostly tolerable in practice: Licensees can
choose any of the Underlying Licenses to follow, and if compliance is
called into question then the technicalities can be examined to ensure
that at least one of the Underlying Licenses is exercised properly.

The additional waiver condition in the old Section 2(d) ("or exercising
only the Default License") must be preserved to ensure that permissions
are not taken away in the new version. This has been split out into
separate sections and explicitly discouraged.